working convolve model
