+++ cifar_10_decision_tree/requirements.txt
